pyproject.toml
[project]
name = "iff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.0",
    "openai>=1.60.0",
    "pyyaml>=6.0.2",
]
